Sliders Logic Complete.

+++ Skin/OldSkinSystem.cs
@@ -28,8 +28,8 @@ namespace MediaPlayer_X_Ark.Skin
 		public Image SliderImage;
 		public Orientation Orientation;
 		public RECT Position;
-		public uint Maximum;
-		public uint Minimum;
+		public int Maximum;
+		public int Minimum;
 	}
 	
 	public struct GraphicCOmponents
@@ -269,8 +269,8 @@ namespace MediaPlayer_X_Ark.Skin
 
 				result.Position.Left = (int)Win32API.GetPrivateProfileInt(section, "-BarX", 0, skinFile);
 				result.Position.Top = (int)Win32API.GetPrivateProfileInt(section, "-BarY", 0, skinFile);
-				result.Minimum = Win32API.GetPrivateProfileInt(section, "-BarMin", 0, skinFile);
-				result.Maximum = Win32API.GetPrivateProfileInt(section, "-BarMax", 0, skinFile);
+				result.Minimum = (int)Win32API.GetPrivateProfileInt(section, "-BarMin", 0, skinFile);
+				result.Maximum = (int)Win32API.GetPrivateProfileInt(section, "-BarMax", 0, skinFile);
 				if (result.Orientation == Orientation.Horizontal)
 				{
 					result.Position.Width = (int)Win32API.GetPrivateProfileInt(section, "-BarAreaX2", 0, skinFile) - result.Position.Left + result.SliderImage.Width;
